requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.7
pandas==2.1.4
openpyxl==3.1.2
celery[redis]==5.3.6
